refactor: use api helpers for frontend requests
