+++ frontend/api.js
@@ -15,12 +15,17 @@ export async function apiGet(path) {
 
 export async function apiPost(path, data) {
     const url = (path.startsWith('/') ? path : `/${path}`);
-    const res = await fetch(url, {
+    const init = {
         method: 'POST',
         credentials: 'include',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(data ?? {}),
-    });
+    };
+    if (data instanceof FormData) {
+        init.body = data;
+    } else {
+        init.headers = { 'Content-Type': 'application/json' };
+        init.body = JSON.stringify(data ?? {});
+    }
+    const res = await fetch(url, init);
     let body = null; try { body = await res.json(); } catch { }
     if (!res.ok) {
         const err = new Error(`HTTP ${res.status}`); err.status = res.status; err.body = body; throw err;
