+++ frontend/components/Profile.vue
@@ -87,7 +87,7 @@
 
 <script setup>
 import { onMounted, ref } from 'vue';
-import { API_BASE } from '../api.js';
+import { apiGet, apiPost } from '../api.js';
 
 const emit = defineEmits(['back', 'logout'])
 const user = ref(null)
@@ -106,14 +106,11 @@ const defaultAvatars = Array.from(
 
 onMounted(async () => {
   try {
-    const res = await fetch(`${API_BASE}/auth/me`, { credentials: 'include' })
-    if (res.ok) {
-      user.value = await res.json()
-      selected.value = user.value.color_palette || 'palette1'
-      document.documentElement.setAttribute('data-theme', selected.value)
-      const res2 = await fetch(`${API_BASE}/games/user/${user.value.user_id}`, { credentials: 'include' })
-      Object.assign(user.value, await res2.json())
-    }
+    user.value = await apiGet('/auth/me')
+    selected.value = user.value.color_palette || 'palette1'
+    document.documentElement.setAttribute('data-theme', selected.value)
+    const stats = await apiGet(`/games/user/${user.value.user_id}`)
+    Object.assign(user.value, stats)
   } catch (err) {
     console.error('Erreur lors du chargement du profil:', err)
   }
@@ -130,15 +127,8 @@ async function onFileChange(e) {
   const form = new FormData()
   form.append('file', file)
   try {
-    const res = await fetch(`${API_BASE}/auth/me/avatar`, {
-      method: 'POST',
-      credentials: 'include',
-      body: form
-    })
-    if (res.ok) {
-      const data = await res.json()
-      if (user.value) user.value.avatar_url = data.avatar_url
-    }
+    const data = await apiPost('/auth/me/avatar', form)
+    if (user.value) user.value.avatar_url = data.avatar_url
   } catch (err) {
     console.error('Erreur upload avatar:', err)
   }
@@ -148,15 +138,8 @@ async function selectAvatar(url) {
   const form = new FormData()
   form.append('choice', url.split('/').pop())
   try {
-    const res = await fetch(`${API_BASE}/auth/me/avatar`, {
-      method: 'POST',
-      credentials: 'include',
-      body: form
-    })
-    if (res.ok) {
-      const data = await res.json()
-      if (user.value) user.value.avatar_url = data.avatar_url
-    }
+    const data = await apiPost('/auth/me/avatar', form)
+    if (user.value) user.value.avatar_url = data.avatar_url
   } catch (err) {
     console.error('Erreur sélection avatar:', err)
   }
@@ -165,12 +148,7 @@ async function selectAvatar(url) {
 async function updatePalette() {
   document.documentElement.setAttribute('data-theme', selected.value)
   try {
-    await fetch(`${API_BASE}/auth/me/palette`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-      body: JSON.stringify({ palette: selected.value })
-    })
+    await apiPost('/auth/me/palette', { palette: selected.value })
     if (user.value) user.value.color_palette = selected.value
   } catch (err) {
     console.error('Erreur mise à jour palette:', err)
@@ -183,10 +161,7 @@ async function deleteAccount() {
     return
   }
   try {
-    await fetch(`${API_BASE}/me/deletion-request`, {
-      method: 'POST',
-      credentials: 'include',
-    })
+    await apiPost('/me/deletion-request')
     emit('logout')
     await window.appAlert('Demande de suppression enregistrée')
   } catch (err) {
