+++ frontend/components/Settings.vue
@@ -13,7 +13,7 @@
 
 <script setup>
 import { onMounted, ref } from 'vue';
-import { API_BASE } from '../api.js';
+import { apiGet, apiPost } from '../api.js';
 
 const emit = defineEmits(['back'])
 const palettes = ['palette1', 'palette2', 'palette3', 'palette4', 'palette5']
@@ -21,12 +21,9 @@ const selected = ref('palette1')
 
 onMounted(async () => {
   try {
-    const res = await fetch(`${API_BASE}/auth/me`, { credentials: 'include' })
-    if (res.ok) {
-      const data = await res.json()
-      selected.value = data.color_palette || 'palette1'
-      document.documentElement.setAttribute('data-theme', selected.value)
-    }
+    const data = await apiGet('/auth/me')
+    selected.value = data.color_palette || 'palette1'
+    document.documentElement.setAttribute('data-theme', selected.value)
   } catch (err) {
     console.error('Erreur chargement palette:', err)
   }
@@ -35,12 +32,7 @@ onMounted(async () => {
 async function updatePalette() {
   document.documentElement.setAttribute('data-theme', selected.value)
   try {
-    await fetch(`${API_BASE}/auth/me/palette`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      credentials: 'include',
-      body: JSON.stringify({ palette: selected.value })
-    })
+    await apiPost('/auth/me/palette', { palette: selected.value })
   } catch (err) {
     console.error('Erreur mise à jour palette:', err)
   }
