+++ frontend/public/authHeartbeat.js
@@ -1,15 +1,12 @@
+import { apiPost } from './api.js'
+
 let timer
 
 function startAuthHeartbeat() {
   stopAuthHeartbeat()
   timer = setInterval(async () => {
     try {
-      const url = `${API_BASE}/auth/refresh`
-      console.log('[authHeartbeat] refreshing', url)
-      await fetch(url, {
-        method: 'POST',
-        credentials: 'include'
-      })
+      await apiPost('/auth/refresh')
     } catch (err) {
       console.error('[authHeartbeat] refresh failed', err)
     }
